feat: order puzzle ids by row_num col
changing the value of the min_moves column in the db might change the order of the rows. then the number assigned to each puzzle in the dropdown list will not be consistent.

row_num is a automatically incrementing integer column. this will represent the number assigned to each puzzle in the dropdown list.