fix: readonly inputs when showing numButtons on desktop
touch devices that have a desktop width (e.g. ipads in a horizontal orientation) should not show the keyboard if numButtons are showing.

the downside is that desktops are no longer able to use the keyboard to input when the numButtons are showing. but this seems reasonable as the default is for numButtons to not be showing on desktop width screens.